Got rid of all usages of PipelineService.findPipelineByCounterOrLabel
